feat: moved voucher instructions steps into a separate folder (PIN-9718)
