+++ src/pages/ConsumerSimulateGetVoucherPage/components/VoucherInstructions/steps/VoucherInstructionsAccessTokenStep.tsx
@@ -5,10 +5,10 @@ import { AUTHORIZATION_SERVER_TOKEN_CREATION_URL, FE_URL } from '@/config/env'
 import { Alert, Divider, Stack, Typography, Link as MUILink } from '@mui/material'
 import { Trans, useTranslation } from 'react-i18next'
 import { useClientKind } from '@/hooks/useClientKind'
-import { CodeSnippetPreview } from './CodeSnippetPreview'
+import { CodeSnippetPreview } from '../CodeSnippetPreview'
 import { InformationContainer } from '@pagopa/interop-fe-commons'
 import { Link } from '@/router'
-import { useVoucherInstructionsContext } from './VoucherInstructionsContext'
+import { useVoucherInstructionsContext } from '../VoucherInstructionsContext'
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
 import { useSearchParams } from 'react-router-dom'
