+++ src/pages/ConsumerSimulateGetVoucherPage/components/VoucherInstructions/steps/VoucherInstructionsClientAssertionStep.tsx
@@ -4,10 +4,10 @@ import { SectionContainer } from '@/components/layout/containers'
 import { Trans, useTranslation } from 'react-i18next'
 import { CLIENT_ASSERTION_JWT_AUDIENCE, FE_URL } from '@/config/env'
 import { useClientKind } from '@/hooks/useClientKind'
-import { CodeSnippetPreview } from './CodeSnippetPreview'
+import { CodeSnippetPreview } from '../CodeSnippetPreview'
 import { StepActions } from '@/components/shared/StepActions'
 import { InformationContainer } from '@pagopa/interop-fe-commons'
-import { useVoucherInstructionsContext } from './VoucherInstructionsContext'
+import { useVoucherInstructionsContext } from '../VoucherInstructionsContext'
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
 import { useSearchParams } from 'react-router-dom'
